requirements.txt
cartopy>=0.21.1
ExifRead>=3.0.0
faiss_cpu>=1.7.3
matplotlib>=3.7.1
numpy>=1.24.2
pandas>=1.5.3
Pillow>=9.4.0
seaborn>=0.12.2
shapely>=2.0.1
torch>=2.0.0
torchvision>=0.15.1
tqdm>=4.65.0
utm>=0.7.0
